Use #dig instead of #[] for obtaining address info
In a tracking event, it is not guaranteed that an address will exist in
the event details.  This is not clear in the Fedex documentation
(https://www.fedex.com/us/developer/downloads/pdf/2019/FedEx_WebServices_DevelopersGuide_v2019.pdf)
but in the WSDL
(https://www.fedex.com/us/developer/downloads/wsdl/2019/standard/TrackService.zip)
the spec is as follows:

```xml
<xs:complexType name="TrackEvent">
  <xs:annotation>
    <xs:documentation>FedEx scanning information about a package.</xs:documentation>
  </xs:annotation>
  <xs:sequence>
    <xs:element name="Address" type="ns:Address" minOccurs="0">
      <xs:annotation>
        <xs:documentation>Address information of the station that is responsible for the scan.</xs:documentation>
      </xs:annotation>
    </xs:element>
```

Therefore, the code must be able to handle the case where no Address
block exists.  This change extracts values with #dig which will return
`nil` if any intermediate step returns `nil` from a Hash/Array.  It
could be asked if there are any repercussions for returning nil values.
There is precedent for events to have an Address block but not contain
any address info.  For example, the "Shipment information sent to Fedex"
event contains an Address block but city, state/province, postal code,
etc. all do not exist and would be set to `nil`.